requirements.txt
accelerate==1.3.0
cupy-cuda12x==13.4.1
datasets==3.6.0
evaluate==0.4.3
lm_eval==0.4.8
torch==2.6.0
transformers==4.52.0